==> Project/2.py <==
import pandas as pd
rollnos = ["CS22M1007",
    "CS22M1001",
    "EC22M1002",
    "ME20B2018",
    "EC21B1044",
    "EC21B1022",
    "ME21B2009",
    "ME21B2008",
    "ME21B2012",
    "ME21B2007",
    "CS21B1043",
    "CS21B1041",
    "MPD191004",
    "CS20B1057",
    "CS21B1052",
    "CS21B1034",
    "CS21B1075",
    "CS21B1033",
    "CS21B1025",
    "CS21B1015",
    "CS21B1003",
    "CS21B1086",
    "CS21B1082",
    "CS21B1088",
    "CS21B2042",
    "CS22B1018",
    "CS20B1045",
    "CS20B1072",
    "CS21B1048",
    "CS21B1087",
    "CS21B1011",
    "CS21B1019",
    "CS21B1083",
    "CS20B1007",
    "EC22B1004",
    "CS20B1040",
    "CS20B1006",
    "CS20B1122",
    "CS20B1128",
    "CS20B1127",
    "CS21B2044",
    "CS20B1047",
    "EC19B1010",
    "EC20B1001",
    "EC20B1015",
    "EC20B1023",
    "EC20B1043",
    "EC20B1101",
    "EC20B1125",
    "EC21B1008",
    "EC21B1028",
    "EC21B1055",
    "EC21B1090",
    "EC21B2043",
    "ME21B2002",
    "ME21B1014",
    "EC21B1026",
    "CS20B1096",
    "ME22B1053",
    "EVD191012",
    "CS21B1053",
    "EC22B1001",
    "ME22M1001",
    "ME23D0002",
    "EC21B1073",
    "CED191029",
    "Ec20B1104",
    "ME20B1008",
    "EC21B1030",
    "ME22B1060",
    "ME22B1046",
    "ME21B2002",
    "ME22B2012",
    "ME22B2011",
    "EC22B1076",
    "EC22B1065",
    "EC22B1125",
    "EC22B1122",
    "EC22B1035",
    "EC21B1027",
    "ME22M1001",
    "ME22B2023",
    "ME22B1047",
    "CS20B1109",
    "CS20B1100",
    "ME22B2025",
    "CS22B1021",
    "MDM20D005",
    "CS22B1047",
    "ME21B1007",
    "EC22B1010",
    "EC21B1075",
    "CS21B2023",
    "ME20B1022",
    "EC20B1051",
    "CS21B2036",
    "EC21B1052",
    "CS22B1024",
    "EC20B1038",
    "CS20B1023",
    "CS22B1072",
    "EC22B1039",
    "EC22B1013",
    "EC21B1015",
    "CS21B1031",
    "ME22B1049",
    "EC22B1069"
    "EC22B1064",
    "EC22B1075"
    "EC22B1057",
    "ID23M1012",
    "CS21B1061",
    "CS21B1056",
    "CS21B1065",
    "CS22B1099",
    "ME22B2010",
    "ME22B2021",
    "EC21B1076",
    "EC21B1072",
    "ME21B1011",
    "EC21B1078",
    "EC21B1090",
    "EC21B1115",
    "EC21B1070",
    "EC21B1066",
    "EC21B1110",
    "EC21B1063",
    "ESD191008",
    "CS21B1114",
    "CS21B1029",
    "CS21B1038",
    "CS21B1030",
    "CS21B1032",
    "CS20B1088",
    "EC22B1015",
    "EC22B1050",
    "CS22M1005",
    "CS21B1085",
    "EC22B1052",
    "MDM20D007",
    "ME20B2018",
    "EC22B1007",
    "EC22B1073",
    "ME21B1042",
    "ME21B1053",
    "CS21B1050",
    "ME20B2017",
    "ME20B1008",
    "EC21B1040",
    "ME22B2027",
    "ME22B2028",
    "ME22B2029",
    "CS22B1043",
    "ME22B2040",
    "CS22B1021",
    "CS22B1024",
    "CS22B2010",
    "CS22B1084",
    "CS22B1089",
    "CS22B1085",
    "EC22B1047",
    "EC22B1042",
    "EC22B1049",
    "ME22B2010",
    "EC22B2021",
    "Me22B2012",
    "ME22B2011",
    "EC2281076",
    "EC22B1065",
    "ID23M1010",
    "ME21D0001",
    "EC20B1021",
    "ID23M1001",
    "ME21B1029",
    "EC22B1031",
    "ID23M1004",
    "CS20B1002",
    "CS20B1006",
    "EC22B1039",
    "EC22B1135",
    "CS22B1012",
    "CS22B1081",
    "EC22B1006",
    "CS22B1009",
    "CS22B1007",
    "ME21B2007",
    "CS21B1041",
    "EC21B1090",
    "EC21B1115",
    "CED191029",
    "EC21B1075",
    "EC21B1065",
    "CS20B1035",
    "EC21B1044",
    "CS21B1054",
    "EC21B1084",
    "EC21B1079",
    "CS21B1101",
    "CS21B2045",
    "CS21B1054",
    "CS21B1048",
    "Cs21B2009",
    "CS21B2007",
    "CS21B2040",
    "EC21B1027",
    "EC21B1004",
    "Cs20B1127",
    "CS20B1006",
    "EC21B1042",
    "CS20B1125",
    "ME20B1022",
    "CS21B2038",
    "ME22B1053",
    "EVD191012",
    "ME20B2037",
    "ME20B2012",
    "ME20B2008",
    "CS21B1005",
    "EC20B1001",
    "EC22B1013",
    "EC22M1005",
    "EC22M1001",
    "CS22B2039",
    "EC22B1005",
    "EC22B1003",
    "CS20B1016",
    "CS22B1028",
    "CS22B1010",
    "CS22B1040",
    "CS22B1051",
    "CS22B1095",
    "EC22B1020",
    "EC22B1045",
    "CS2081107",
    "CS20B1066",
    "Ec22B1026",
    "CS21B1038",
    "CS21B2014",
    "CS21B2015",
    "CS21B2012",
    "CS2282015",
    "Cs21B1046",
    "CS22B1057",
    "ESD191001",
    "CS21B1045",
    "CS21B2034",
    "CS21B2033",
    "ME20B1033",
    "CS20B1057",
    "ME21B2013",
    "ME21B2019",
    "ME21B2016",
    "ME21B1055",
    "ESD191024",
    "ME22B1014",
    "ME22B1044",
    "EC21B1005",
    "EC22B1117",
    "MAT180004",
    "EC22B1007",
    "EC21B1081",
    "EC20B1051",
    "EC21B1053",
    "EC21B1052",
    "EC21B1037",
    "EC21B1032",
    "Ec21B1048",
    "EC22B1073",
    "EC21B1059",
    "EC21B1050",
    "ME21B1044",
    "ME20B1055",
    "ME21B2002",
    "ME20B2014",
    "EC20B1109",
    "CS20B1123",
    "CS21B1067",
    "CS2181063",
    "EC22B1042",
    "ME21B1072",
    "CS21B1057",
    "ME21B1066",
    "ME21B1070",
    "ME21B1075",
    "ME21B1076",
    "ME22B1075",
    "CS22B2010",
    "EDM20D013",
    "ME22B1048",
    "CS22B1036",
    "EC22B1023",
    "EC22B1048",
    "CS22B1066",
    "EC22B1099",
    "EC22B1092",
    "EC22B1093",
    "ME21B2015",
    'CS21B1034',
    'CS21B1015',
    'CS21B1086',
    'CS21B1013',
    'CS22B1003',
    'CS22B1002',
    'CS22B1001',
    'CS21B2043',
    'CS21B2042',
    'Cs21B1010',
    'CS21B2025',
    'CS21B1062',
    'CS21B2024',
    'CS22B2019',
    'ME22B1046',
    'ME22B2043',
    'ME22B1043',
    'ME22B2020',
    'CS22B1079',
    "ME21B2009",
    "CS21B1087", 
    "ME21B2022", 
    "EC22B1004", 
    "CS22B1047", 
    "ME22D0002", 
    "ME22D0004", 
    "ME2181057",      
    "ME21B1061", 
    "EC21B1057", 
    "CS21B1080", 
    "CS22B1098",
    "CS22B1092",
    "CS22B2055", 
    "CS22B2051",
    "CS22B2021", 
    "EC22B1027", 
    "CS22B1088",
    "CS22B1090", 
    "EC20B1020", 
    "ME20B1024", 
    "EC22B1050", 
    "EC21B1071", 
    "EC22B1109",
    "EC22B1104",
    "EC22B1110", 
    "CS22B1025",
    "ME22B1072", 
    "ME22B1073", 
    "EC21B1107", 
    "ME22B2006",
    "ME22B2008", 
    "ME21B1022", 
    "CS21B2019", 
    "CS21B2029",
    "CS21B1009", 
    "CS21B1022", 
    "CS21B2030", 
    "CS21B2004", 
    "CS21B1051", 
    "CS22B2012", 
    "EC22B1062", 
    "CS22B1067",
    "CS22B1060", 
    "EC22B1056", 
    "EC22B1058",
    "CS20B1012", 
    "CS21B1053", 
    "CS20B1122", 
    "CS20B1128", 
    "EC22B1100", 
    "EC22B1036", 
    "CS22B2054", 
    "CS20B1042", 
    "CS22B2023", 
    "CS22B2020", 
    "EC22B1088", 
    "CS22B2031", 
    "ME20B1001", 
    "EC20B1017", 
    "CS21B1007", 
    "EC22B1113", 
    "ECS22B1093",
    "ME21B2027", 
    "EC22B1040", 
    "CS22B1050", 
    "CS22B1097", 
    "CS20B1116", 
    "EC22B1055", 
    "CS22B1070", 
    "CS22B1097", 
    "CS22B1083", 
    "ECS22B1093", 
    "ME21B1027", 
    "ME21B1069", 
    "ME22B1026", 
    "ME21B1026", 
    "CS22B1097", 
    "CS22B2026",
]

df = pd.DataFrame({'Roll No': rollnos})
df.to_csv('RollNos.csv', index=False)
